pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-guard"
version = "0.1.0"
description = "Lightweight prompt injection detection for LLM applications"
requires-python = ">=3.11"
license = "MIT"

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0", "httpx>=0.27.0", "pyyaml>=6.0"]

[tool.setuptools.packages.find]
where = ["src"]
